# [chore][exporter][batcher] Improve exporter.request merge splitting performance by caching items count